Move get_hypervisor_type method from utils to common
