+++ test/integration/smoke/test_iso.py
@@ -19,12 +19,15 @@
 # Import Local Modules
 from marvin.cloudstackTestCase import cloudstackTestCase, unittest
 from marvin.cloudstackAPI import listZones, updateIso, extractIso, updateIsoPermissions, copyIso, deleteIso
-from marvin.lib.utils import cleanup_resources, random_gen, get_hypervisor_type,validateList
+from marvin.lib.utils import cleanup_resources, random_gen, validateList
 from marvin.lib.base import Account, Iso
-from marvin.lib.common import (get_domain,
-                               get_zone,
-                               list_isos,
-                               list_os_types)
+from marvin.lib.common import (
+    get_domain,
+    get_zone,
+    list_isos,
+    list_os_types,
+    get_hypervisor_type
+)
 from nose.plugins.attrib import attr
 from marvin.codes import PASS
 import urllib
